+++ packages/theme_tailor/lib/src/generator/generator_for_annotated_class.dart
@@ -1,4 +1,4 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:theme_tailor/src/generator/generator_annotation_matcher.dart';
@@ -10,12 +10,12 @@ abstract class GeneratorForAnnotatedClass<TLibraryData, TAnnotationData, TData, 
   @override
   void generateToBuffer(
     StringBuffer buffer,
-    Element element,
+    Element2 element,
     ConstantReader annotation,
   ) {
     final classElement = ensureClassElement(element);
     final data = parseData(
-      parseLibraryData(classElement.library, classElement),
+      parseLibraryData(classElement.library2, classElement),
       parseAnnotation(annotation),
       classElement,
     );
@@ -25,7 +25,7 @@ abstract class GeneratorForAnnotatedClass<TLibraryData, TAnnotationData, TData, 
 
   @override
   String generateForAnnotatedElement(
-    Element element,
+    Element2 element,
     ConstantReader annotation,
     BuildStep buildStep,
   ) {
@@ -34,16 +34,16 @@ abstract class GeneratorForAnnotatedClass<TLibraryData, TAnnotationData, TData, 
     return buffer.toString();
   }
 
-  TLibraryData parseLibraryData(LibraryElement library, ClassElement element);
+  TLibraryData parseLibraryData(LibraryElement2 library, ClassElement2 element);
 
   TAnnotationData parseAnnotation(ConstantReader annotation);
 
-  ClassElement ensureClassElement(Element element);
+  ClassElement2 ensureClassElement(Element2 element);
 
   TData parseData(
     TLibraryData libraryData,
     TAnnotationData annotationData,
-    ClassElement element,
+    ClassElement2 element,
   );
 
   void generateForData(StringBuffer buffer, TData data);