+++ packages/theme_tailor/lib/src/util/extension/constructor_element_extension.dart
@@ -1,9 +1,0 @@
-import 'package:analyzer/dart/element/element.dart';
-
-extension ConstructorElementExtension on ConstructorElement {
-  bool hasAllParametersNamed(Set<String> params) {
-    return params.containsAll(_parameterNames);
-  }
-
-  Iterable<String> get _parameterNames => parameters.map((e) => e.name);
-}