+++ packages/theme_tailor/lib/src/util/extension/element_annotation_extension.dart
@@ -1,4 +1,4 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:theme_tailor_annotation/theme_tailor_annotation.dart';
 
@@ -8,6 +8,7 @@ extension ElementAnnotationExtension on ElementAnnotation {
   }
 
   bool get isTailorThemeExtension {
-    return TypeChecker.fromRuntime(themeExtension.runtimeType).isAssignableFrom(computeConstantValue()!.type!.element!);
+    return TypeChecker.fromRuntime(themeExtension.runtimeType)
+        .isAssignableFrom(computeConstantValue()!.type!.element3!);
   }
 }