+++ packages/theme_tailor_toolbox/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 3.1.0
+- Require Dart 3.8.0
+- Use the same versioning for other theme_tailor packages
+
 # 1.0.6
 - Added LinearGradientEncoder
 - Added SweepGradientEncoder